Switched contract to use longs. Added out of range test. Renamed some tests.

+++ PrimeNumbersSignalR/PrimeNumbersSignalR.Tests/Hubs/NumberGenerationHub/INumberGenerationClientContract.cs
@@ -8,7 +8,7 @@ namespace PrimeNumbersSignalR.Hubs.Tests
 {
     public interface INumberGenerationClientContract
     {
-        void generatedNumbers(IList<int> generatedNumbers);
+        void generatedNumbers(IList<long> generatedNumbers);
 
         void generatedNumbers(Exception generationException);
     }
